+++ src/core/QCheck2.mli
@@ -280,16 +280,32 @@ module Gen : sig
       Shrinks towards [0.].
   *)
 
+  val float_pos : float t
+  (** Generates positive floating point numbers ([0.] included).
+
+      Shrinks towards [0.].
+  *)
+
+  val float_neg : float t
+  (** Generates negative floating point numbers. ([-0.] included).
+
+      Shrinks towards [-0.].
+  *)
+
   val pfloat : float t
   (** Generates positive floating point numbers ([0.] included).
 
       Shrinks towards [0.].
+
+      @deprecated use {!float_pos} instead.
   *)
 
   val nfloat : float t
   (** Generates negative floating point numbers. ([-0.] included).
 
       Shrinks towards [-0.].
+
+      @deprecated use {!float_neg} instead.
   *)
 
   val char : char t
@@ -541,11 +557,16 @@ module Gen : sig
 
       @since 0.11 *)
 
-  val exponential : float -> float t
-  (** [exponential m] generates floating-point numbers following an exponential
+  val float_exp : float -> float t
+  (** [float_exp m] generates floating-point numbers following an exponential
       distribution with a mean of [m].
 
       @raise Invalid_argument if [m] is NaN.
+
+      @since NEXT_RELEASE *)
+
+  val exponential : float -> float t
+  (** Synonym for {!float_exp}.
 
       @since 0.23 *)