+++ test/core/QCheck_tests.ml
@@ -617,33 +617,33 @@ module Shrink = struct
     Test.make ~name:"float_range -10. -1. >= -.pi" ~count:1000
       (float_range (-10.) (-1.)) (fun f -> f >= -.Float.pi)
 
-  let pos_float_lt_pi =
-    Test.make ~name:"pos_float < Float.pi" ~count:1000
-      pos_float (fun f -> f < Float.pi)
+  let float_pos_lt_pi =
+    Test.make ~name:"float_pos < Float.pi" ~count:1000
+      float_pos (fun f -> f < Float.pi)
 
-  let pos_float_not_nan =
-    Test.make ~name:"pos_float is not nan" ~count:10_000
-      pos_float (fun f -> not (Float.is_nan f))
+  let float_pos_not_nan =
+    Test.make ~name:"float_pos is not nan" ~count:10_000
+      float_pos (fun f -> not (Float.is_nan f))
 
-  let pos_float_not_infinite =
-    Test.make ~name:"pos_float is not infinity" ~count:10_000
-      pos_float (fun f -> not (Float.is_infinite f))
+  let float_pos_not_infinite =
+    Test.make ~name:"float_pos is not infinity" ~count:10_000
+      float_pos (fun f -> not (Float.is_infinite f))
 
-  let neg_float_gt_mpi =
-    Test.make ~name:"neg_float > Float.pi" ~count:1000
-      neg_float (fun f -> f > -.Float.pi)
+  let float_neg_gt_mpi =
+    Test.make ~name:"float_neg > Float.pi" ~count:1000
+      float_neg (fun f -> f > -.Float.pi)
 
-  let neg_float_not_nan =
-    Test.make ~name:"neg_float is not nan" ~count:10_000
-      neg_float (fun f -> not (Float.is_nan f))
+  let float_neg_not_nan =
+    Test.make ~name:"float_neg is not nan" ~count:10_000
+      float_neg (fun f -> not (Float.is_nan f))
 
-  let neg_float_not_infinite =
-    Test.make ~name:"neg_float is not infinity" ~count:10_000
-      neg_float (fun f -> not (Float.is_infinite f))
+  let float_neg_not_infinite =
+    Test.make ~name:"float_neg is not infinity" ~count:10_000
+      float_neg (fun f -> not (Float.is_infinite f))
 
-  let exponential_10_lt_pi =
-    Test.make ~name:"exponential 10. < Float.pi" ~count:1000
-      (exponential 10.) (fun f -> f < Float.pi)
+  let float_exp_10_lt_pi =
+    Test.make ~name:"float_exp 10. < Float.pi" ~count:1000
+      (float_exp 10.) (fun f -> f < Float.pi)
 
   let char_is_never_abcdef =
     Test.make ~name:"char never produces 'abcdef'" ~count:1000
@@ -927,13 +927,13 @@ module Shrink = struct
     float_range_1_10_leq_pi;
     float_range_m10_10_square_leq_2;
     float_range_m10_m1_geq_mpi;
-    pos_float_lt_pi;
-    pos_float_not_nan;
-    pos_float_not_infinite;
-    neg_float_gt_mpi;
-    neg_float_not_nan;
-    neg_float_not_infinite;
-    exponential_10_lt_pi;
+    float_pos_lt_pi;
+    float_pos_not_nan;
+    float_pos_not_infinite;
+    float_neg_gt_mpi;
+    float_neg_not_nan;
+    float_neg_not_infinite;
+    float_exp_10_lt_pi;
     char_is_never_abcdef;
     printable_is_never_sign;
     numeral_is_never_less_5;
@@ -1225,10 +1225,10 @@ module Stats = struct
       Test.make ~name:"int64 upper dist"     ~count:10000 (add_stat64 (fun i -> Int64.shift_right_logical i 48) int64) (fun _ -> true);
     ]
 
-  let exponential_tests =
+  let float_exp_tests =
     let float_dist = ("dist",int_of_float) in
-    [ Test.make ~name:"exponential 10. dist" ~count:5_000 (add_stat float_dist (exponential 10.)) (fun _ -> true);
-      Test.make ~name:"exponential -10. dist" ~count:5_000 (add_stat float_dist (exponential (-10.))) (fun _ -> true);
+    [ Test.make ~name:"float_exp 10. dist" ~count:5_000 (add_stat float_dist (float_exp 10.)) (fun _ -> true);
+      Test.make ~name:"float_exp -10. dist" ~count:5_000 (add_stat float_dist (float_exp (-10.))) (fun _ -> true);
     ]
 
   let float_tests = (* Float.frexp nan (and infinity) is undefined and may return a 32766 exponent on Alpine *)
@@ -1276,6 +1276,6 @@ module Stats = struct
     @ array_len_tests
     @ int_dist_tests
     @ int_32_64_dist_tests
-    @ exponential_tests
+    @ float_exp_tests
     @ float_tests
 end