+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## NEXT RELEASE (202?-??-??)
 
 - Remove deprecated function generators `QCheck.{fun1_unsafe,fun2_unsafe}`
+- Renamed `float` generator combinators:
+  - Add `QCheck.Gen.{float_pos, float_neg}` and deprecate `QCheck.Gen.{pfloat, nfloat}`
+  - Add `QCheck.{float_pos, float_neg}` and deprecate `QCheck.{pos_float, neg_float}`
+  - Add `QCheck2.{float_pos, float_neg}` and deprecate `QCheck2.{pfloat, nfloat}`
+  - Add `QCheck.(--.)` as a synonym for `QCheck.float_range` for consistency
+  - Add `float_exp` to `QCheck.Gen`, `QCheck`, and `QCheck2.Gen` and make
+    the existing `exponential` binding a synonym for it
 
 ## 0.27 (2025-10-31)